# Optimize ignited-cells datastructure