--- a/.gitignore
+++ b/.gitignore
@@ -9,3 +9,4 @@ org/auto/
 org/missfont.log
 target/
 server.edn
+scratch.clj