--- a/src/gridfire/fire_spread.clj
+++ b/src/gridfire/fire_spread.clj
@@ -236,18 +236,17 @@
 (defn generate-ignited-cells
   [constants fire-spread-matrix cells]
   (when (seq cells)
-    (persistent!
-     (reduce (fn [ignited-cells cell]
-               (apply conj
-                      ignited-cells
-                      (compute-neighborhood-fire-spread-rates! constants
-                                                               fire-spread-matrix
-                                                               cell
-                                                               nil
-                                                               0.0
-                                                               0.0)))
-             (transient [])
-             cells))))
+    (reduce (fn [ignited-cells cell]
+              (apply conj
+                     ignited-cells
+                     (compute-neighborhood-fire-spread-rates! constants
+                                                              fire-spread-matrix
+                                                              cell
+                                                              nil
+                                                              0.0
+                                                              0.0)))
+            []
+            cells)))
 
 (defn identify-spot-ignition-events
   [global-clock spot-ignitions]