## This PR:

- converts the representation of ignited-cells 
  - Old structure: map of source-cell -> list of neighborhood trajectories
  - New structure: list of trajectories

This gave us a 15% improvement in runtime

Optimized 
```
pId                           nCalls        Min        Max       Mean   MAD      Clock  Total

:run-fire-spread                  10     5.02ms     5.80s      2.55s   ±99%    25.53s     96%
:update-ignited-cells          6,011     4.30μs    10.64ms     3.02ms  ±42%    18.14s     68%
:identify-ignition-events      6,011     1.10μs     4.52ms   602.58μs  ±45%     3.62s     14%
:update-matrices               6,011   600.00ns     4.58ms   246.63μs  ±56%     1.48s      6%
:calculate-dt                  6,011   800.00ns     1.03ms   244.55μs  ±52%     1.47s      6%

Accounted                                                                      50.25s    190%
Clock                                                                          26.50s    100%
```

Not Optimized
```
04/16 12:05:50 :run-simulation,
pId                           nCalls        Min        Max       Mean   MAD      Clock  Total

:run-fire-spread                  10     5.35ms     7.00s      3.06s   ±99%    30.64s     97%
:update-ignited-cells          6,011     4.10μs    10.96ms     3.36ms  ±42%    20.19s     64%
:identify-ignition-events      6,011     1.70μs     4.85ms   847.80μs  ±47%     5.10s     16%
:calculate-dt                  6,011     1.90μs     2.01ms   506.71μs  ±50%     3.05s     10%
:update-matrices               6,011   600.00ns     4.59ms   248.79μs  ±57%     1.50s      5%

Accounted                                                                       1.01m    191%
Clock                                                                          31.57s    100%
```